+++ Meraki.Api/Data/OrganizationWirelessRadioRfHealthOverviewByNetworkByIntervalItemNetwork.cs
@@ -1,0 +1,22 @@
+namespace Meraki.Api.Data;
+
+/// <summary>
+/// Organization Wireless Radio Rf Health Overview By Network By Interval Item Network
+/// </summary>
+[DataContract]
+public class OrganizationWirelessRadioRfHealthOverviewByNetworkByIntervalItemNetwork
+{
+	/// <summary>
+	/// The network ID.
+	/// </summary>
+	[ApiAccess(ApiAccess.Read)]
+	[DataMember(Name = "id")]
+	public string Id { get; set; } = string.Empty;
+
+	/// <summary>
+	/// The network name.
+	/// </summary>
+	[ApiAccess(ApiAccess.Read)]
+	[DataMember(Name = "name")]
+	public string Name { get; set; } = string.Empty;
+}